requirements.txt
pandas==2.2.2
numpy==1.26.4
pydantic==2.7.4
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
streamlit==1.36.0
pytest==8.2.1
ruff==0.5.4
isort==5.13.2
pyright==1.1.374
